Fixed initialization part
Fixed initialization part: If we try to pass "this" pointer in CreateWindowEx inside the constructor of base class, we are trying to pass a pointer to a non-fully constructed object , in this case AppWindow, our derived class! In particular the vtable is not fully constrcuted, so, in the moment we try to call virtual method like onCreate inside Window class, simply the virtual methods of AppWindow will be ignored. and will be called only the Window methods instead. The solution is to move the initialization part inside a suitable method, like init() , and call it after the execution of all the constructors, in this way we are sure that the object will be fully constructed.